Refactor hard-coded maximum number of characters value
The maximum number of characters permitted in a status message is a hard-coded number in the onTextChanged event handler function. It would be better if this were a value that could be changed via a resource value.

Create a new constants.xml resource file
Create a new integer resource for the maximum character permitted